test(appeals): added test cases for partial cost decisions
consolidated other spec files into 'Issue Decision'
