+++ appeals/e2e/cypress/support/happyPathHelper.js
@@ -238,7 +238,7 @@ export const happyPathHelper = {
 		caseDetailsPage.validateBannerMessage('Success', 'Site visit set up');
 	},
 
-	issueDecision(caseRef, decision) {
+	issueDecision(caseRef, decision, appellantCostsBool = true, lpaCostsBool = true) {
 		caseDetailsPage.clickIssueDecision();
 		caseDetailsPage.selectRadioButtonByValue(caseDetailsPage.exactMatch(decision));
 		caseDetailsPage.clickButtonByText('Continue');
@@ -246,16 +246,26 @@ export const happyPathHelper = {
 		caseDetailsPage.clickButtonByText('Continue');
 
 		//Appellant costs
-		caseDetailsPage.selectRadioButtonByValue('Yes');
-		caseDetailsPage.clickButtonByText('Continue');
-		fileUploader.uploadFiles(sampleFiles.pdf);
-		caseDetailsPage.clickButtonByText('Continue');
+		if (appellantCostsBool) {
+			caseDetailsPage.selectRadioButtonByValue('Yes');
+			caseDetailsPage.clickButtonByText('Continue');
+			fileUploader.uploadFiles(sampleFiles.pdf);
+			caseDetailsPage.clickButtonByText('Continue');
+		} else {
+			caseDetailsPage.selectRadioButtonByValue('No');
+			caseDetailsPage.clickButtonByText('Continue');
+		}
 
 		//LPA costs
-		caseDetailsPage.selectRadioButtonByValue('Yes');
-		caseDetailsPage.clickButtonByText('Continue');
-		fileUploader.uploadFiles(sampleFiles.pdf);
-		caseDetailsPage.clickButtonByText('Continue');
+		if (lpaCostsBool) {
+			caseDetailsPage.selectRadioButtonByValue('Yes');
+			caseDetailsPage.clickButtonByText('Continue');
+			fileUploader.uploadFiles(sampleFiles.pdf);
+			caseDetailsPage.clickButtonByText('Continue');
+		} else {
+			caseDetailsPage.selectRadioButtonByValue('No');
+			caseDetailsPage.clickButtonByText('Continue');
+		}
 
 		//CYA
 		caseDetailsPage.clickButtonByText('Issue decision');
@@ -263,6 +273,42 @@ export const happyPathHelper = {
 		//Banner & inset text
 		caseDetailsPage.validateBannerMessage('Success', 'Decision issued');
 		caseDetailsPage.checkStatusOfCase('Complete', 0);
+	},
+
+	issueInvalidDecision(caseRef, appellantCostsBool = true, lpaCostsBool = true) {
+		caseDetailsPage.clickIssueDecision();
+
+		//issue decision
+		caseDetailsPage.selectRadioButtonByValue('Invalid');
+		caseDetailsPage.fillTextArea('The appeal is invalid because of X reason', 0);
+		caseDetailsPage.clickButtonByText('Continue');
+
+		//Appellant costs
+		if (appellantCostsBool) {
+			caseDetailsPage.selectRadioButtonByValue('Yes');
+			caseDetailsPage.clickButtonByText('Continue');
+			fileUploader.uploadFiles(sampleFiles.pdf);
+			caseDetailsPage.clickButtonByText('Continue');
+		} else {
+			caseDetailsPage.selectRadioButtonByValue('No');
+			caseDetailsPage.clickButtonByText('Continue');
+		}
+
+		//LPA costs
+		if (lpaCostsBool) {
+			caseDetailsPage.selectRadioButtonByValue('Yes');
+			caseDetailsPage.clickButtonByText('Continue');
+			fileUploader.uploadFiles(sampleFiles.pdf);
+			caseDetailsPage.clickButtonByText('Continue');
+		} else {
+			caseDetailsPage.selectRadioButtonByValue('No');
+			caseDetailsPage.clickButtonByText('Continue');
+		}
+
+		//CYA
+		caseDetailsPage.clickButtonByText('Issue decision');
+		caseDetailsPage.validateBannerMessage('Success', 'Appeal marked as invalid');
+		caseDetailsPage.checkStatusOfCase('Invalid', 0);
 	},
 
 	issueDecisionWithoutCosts(caseRef, decision) {
